+++ CHANGELOG.md
@@ -7,7 +7,14 @@ with regard to its JavaScript and TypeScript bindings' APIs and the Ethereum ABI
 
 ## [Unreleased]
 
+### Added
+- Burning capability to Radicle Token
+
+### Changed
 - Replaced vesting contract with one adapted from Melonport
+
+### Removed
+- Rad token
 
 ## [0.1.0] - 2020-12-01
 ### Added